Fixes update capability
Updates need a Menagerie UUID to identify devices; previously there was
no mechanism to send this.

It has been added as a field on the Update
tab, and will be partially matched when sent to the server (i.e. '57eed'
could be used to update '57eed8d62ad9708a2d1c6a42').